Refactor, create base testcase for bundle factory.
