+++ opengever/bundle/tests/test_bundle_factory_xlsx.py
@@ -1,67 +1,12 @@
 from opengever.bundle.factory import BundleFactory
 from opengever.bundle.factory import parse_args
 from opengever.bundle.tests import assets
-from tempfile import mkdtemp
-from unittest import TestCase
+from opengever.bundle.tests.base import BaseTestOggBundleFactory
 import json
 import os
-import shutil
 
 
-class TestOggBundleFactoryXLSX(TestCase):
-
-    def setUp(self):
-        super(TestOggBundleFactoryXLSX, self).setUp()
-        self.tempdir = mkdtemp()
-
-    def tearDown(self):
-        shutil.rmtree(self.tempdir)
-        super(TestOggBundleFactoryXLSX, self).tearDown()
-
-    def find_item_by_title(self, items, title):
-        for item in items:
-            if item.get('title') or item.get('title_de') == title:
-                return item
-        return None
-
-    def find_items_by_parent_guid(self, items, value):
-        matching_items = []
-        for item in items:
-            if item.get('parent_guid') == value:
-                matching_items.append(item)
-        return matching_items
-
-    def assert_reporoot_default_properties(self, reporoot, user_group):
-        # Title should be stored with the title_de (and not title) key
-        self.assertIn('title_de', reporoot)
-        self.assertNotIn('title', reporoot)
-        self.assertTrue(reporoot.get('title_de'))
-
-        self.assertEqual('repositoryroot-state-active', reporoot['review_state'],
-                         msg="Review state should be set to active")
-
-        # Permissions are set on the repository root
-        self.assertDictEqual({u'read': [user_group],
-                              u'edit': [user_group],
-                              u'add': [user_group],
-                              u'close': [],
-                              u'reactivate': []},
-                             reporoot['_permissions'])
-
-        self.assertFalse(reporoot.get(u'parent_guid') or reporoot.get(u'parent_reference'),
-                         msg="A reporoot cannot define a parent")
-
-    def assert_repofolder_default_properties(self, repofolder):
-        # Title should be stored with the title_de (and not title) key
-        self.assertIn('title_de', repofolder)
-        self.assertNotIn('title', repofolder)
-        self.assertTrue(repofolder.get('title_de'))
-
-        self.assertEqual('repositoryfolder-state-active', repofolder['review_state'],
-                         msg="Review state should be set to active")
-
-        self.assertTrue(repofolder.get(u'parent_guid') or repofolder.get(u'parent_reference'),
-                        msg="A repofolder always needs to define its parent")
+class TestOggBundleFactoryXLSX(BaseTestOggBundleFactory):
 
     def test_xlsx_bundle_factory(self):
         input_path = assets.get_path('basic_repository.xlsx')
